Add a link to the diff on GitHub
I'm one of those young people that is used to look at diffs in the
GitHub interface. The cfbot repo is the easiest way I've found for me to
look at diffs. I even use it as a remote for my local postgres checkout,
so that I can easily check out a certain commitfest entry locally.

This PR adds a link to the github diff for a commitfest entry.

NOTE: this patch is untested, since I don't know how to run this cfbot
repo locally.