Don't squash patches into one big commit
As I explained in #17 the cfbot repo is the easiest way for me to take a
quick look at postgres patches in a decent diff interface, as well as
having an easy way of checking them out locally. Instead of squashing
the patches together it's much nicer if they are committed separately.

The suggested changes try `git am` and if that fails it falls back to
the previous `gpatch` command.

NOTE: Again this is untested code, since I don't know how to get cfbot
running locally.