Don't squash patches into one big commit
As I explained in #17 the cfbot repo is the easiest way for me to take a
quick look at postgres patches in a decent diff interface, as well as
having an easy way of checking them out locally. Instead of squashing
the patches together it's much nicer if they are committed separately.

The suggested changes `git mailinfo` to extract to extract commit info
in case the patch was created by `git format-patch`. It explicitely
doesn't use `git am` to apply the patch, because that chokes on many
more merge conflicts as regular `patch`/`gpatch` does.

P.S. This starts using <<'EOF' instead of <<EOF to not have to escape
dollar signs all the time in the `apply-patches.sh` script.