Add a link to the diff on GitHub
I'm one of those young people that is used to look at diffs in the
GitHub interface. The cfbot repo is the easiest way I've found for me to
look at diffs of a commitfest entry. I even use it as a remote for my
local postgres checkout, so that I can easily check out a certain
commitfest entry locally.

This PR adds a link to the GitHub diff for a commitfest entry.