Time out more promptly, and tell the CF app.
Previously we'd run a timeout job daily.  That wasn't enough and could
jam things up for a long time if the macOS CI machine was down (though a
new Mac is joining the pool soon to reduce that risk a lot).  Now we'll
do that as soon as 1 hour is reached, in the same code path that polls
for task statuses and also does the testing->finished transition.

We also forgot to enqueue a message to tell the CF app about branch
timeouts.  That's now fixed by putting the state change in the same
code path that does all the right things.

Now if cfbot gives up on a branch, it won't rate-limit itself to prevent
new branches, and the CF will know straight away.  That might have other
consequences if Cirrus is still waiting...  Jelte suggests actually
cancelling the build via the Cirrus API in that case, but right now we
don't have code in place to make authenticated API calls.  Something to
look into.