Receive Cirrus webhook POSTs on CI status changes.
For now this just creates cirrus-task-update entries in work_queue which
are ignored/deleted because there is no handler yet.  Needs some
thought, in terms of reliability and timing, since Cirrus won't call us
again on transient failure.  So we probably still need to keep polling,
but with some more hacking we can at least process changes quickly most
of the time...

This URL has already been registered with Cirrus, so the wsgi endpoint
is called in production.  If changed and deployed, the code can be reloaded
with:

$ sudo supervisorctl restart cfbot_api