Switch to event-based polling of Cirrus.
Change the work_queue rows inserted by for Cirrus webook POSTs to
poll-cirrus-branch, and the key to just the commit ID.  Create a handler
that will then run our pre-existing polling code to call Cirrus's API to
find the status of all tasks.  Now that is the primary way that we poll
Cirrus, ie after receiving a poke.

Since we receive a poke for every task that changes and we're only
considering the commit ID, not the task, introduce a way to enqueue
work_queue jobs without creating duplicates.

Since we might miss a poke for whatever reason, also periodically poll
any branches that haven't seen any tast status changes in 5 minutes,
so that we don't get stuck due to lost webhooks.

It's a bit messy that we are using branch IDs in some places and commit
IDs in others, meanwhile Cirrus has a separate build ID that we aren't
using at all -- in theory there could be more than one build for the
same commit -- but that's a pre-existing problem that should eventually
be improved...