Introduce build table from Cirrus's model.
Previously we didn't model "builds" directly, and we'd get quite
confused if Cirrus auto-restart a build because we used only the commit
ID to connect tasks to branches.  This is the beginning of correcting
that historical error.

Now that we have Cirrus's build ID and status, we can simplify the code
that figured out when a branch should be considered 'finished'.

TODO: Put back the backoff code, in tidier form...

Several columns can probably be dropped, with more study on which
queries will break.

WIP