Connect branches to builds directly.
Previously we'd get very confused if Cirrus decided to cancel and
restart a build, since we were joining by commit IDs only.  Earlier
commits added a build table and a task.build_id column.  This adds a
branch.build_id column to record which build we consider the "current"
build for a given branch.  It is kept up to date with some simple
heuristics.  The timeout logic might potentially need some adjustment...

The build_id property is added to messages posted to the cfapp (it
should just ignore unknown dictionary elements), and used to select the
current interest tasks for a given branch instead of the commit ID.

Also skip post-status-update work for tasks that are not connected to a
branch, since the cfapp certainly doesn't want to hear about those (ie
master branch etc).