Add statistics-based polling of tasks.
Poll a build every minutes aftera task status has been stuck in a
running state for longer than 95% of successful builds for the relevant
references branch and task name.

So now there are three separate things that we might poll:

 - branch records that don't have build IDs yet (which we have to link
   up only after we hear from Cirrus)
 - build records that exceed 95% of recent successful times spent in
   that state
 - ditto for task records

In all three cases it's the same basic operation, poll-stale-build, but
with three different reasons to tigger it.

This is a bit of a change from before when we'd only poll for branches
in state 'testing'.  Now, we'll mark the branch 'timeout' after we've
given up waiting for the change so that we can create new branches, but
we'll still continue to poll the stuck task.  That seems OK because
Cirrus will eventually bring it to a final state, right?  It will fill
the log up with spam though.

We could fix that by using the Cirrus API to proactively cancel builds
that we've given up on.  To investigate...

Some basic indexing to find running tasks and builds quickly, but more
research needed on that as data size increases.