Remove bogus polling reason.
Previously I assumed that a build in final status must have all its
tasks in a final status, so if that wasn't true I scheduled a poll of
the build.  That was bogus, because EXECUTING->FAILED happens when the
first task fails, while others are expected to be running still.  That's
OK, because the new periodic poll_stale_tasks() will eventually trigger
a poll when they appear to be running for a statistically unlikely long
time.  Much better, getting simpler...