Use three standard deviations instead of two.
In an attempt to run almost entirely on webhooks without polling, only
poll if the time exceeds 99.7% (3 sigma) of recently observed times.
Not sure if this is a good tradeoff, but now that we can handle common
slightly out-of-order state transitions, we should almost always be able
to count on webhooks, so it was a bit annoying to see occasional polling
happening when nothing was actually wrong.