Drop some obsolete columns from task table.
This schema has gradually evolved to fit Cirrus's model (as you might
guess from the primnary key of this table being near the end...),  It's
time to tidy up some things that don't make sense anymore.
commitfest_id and submission_id had become redundant in the 'task'
table, since you can use the associaed build and branch tables.  So help
me Codd.

commit_id is also redundant, since it can be reached by build_id, but
we'd either have to backfill the new build table or wait for the
buildless tasks to age out of our 90 day window and the build_id column
NOT NULL first, so I'm not ready to drop that one yet.