Rename column task_status_history.timestamp column.
"sent" was a confusing name.  This is the timestamp from Cirrus, not the
time we sent it anywhere.  We also have "received", when we heard about
it.  I suppose "sent" might be interesting as a name to record the time
we sent it to the cfapp, but that's not currently done.