Fix lost build row updates.
Sometimes poll_stale_build() has to insert a build row with missing
data, assuming it would be updated later in process_new_build_status().
That didn't actually happen, so we could finish up with some rows that
had NULL columns.

It would be nicer if those columns were NOT NULL and we didn't have to
insert a sort of "dummy" row, but that's needed for serialisation.  We
could lock something else instead...